seperated dominant_colors package
purpose: loading KMeans takes long time, by seperating the imports it should benefit the loading speed.
