+++ adapters/cv_image.py
@@ -1,7 +1,7 @@
 from cv2 import imread, imshow, waitKey, destroyAllWindows
 from numpy import copy as np_copy, array as np_array, uint8
 from core.image import Image
-from utils.dominant_colors import get_cluster
+from utils.cluster import get_cluster
 
 class OpenCVImage(Image):
   def __init__(self, path):
